Add files via upload
This project was developed with an objective to bring in innovation and improvements in existing ECG machine and the way an ECG output is analysed. Usual practice in interpreting an ECG is analyse ECG waveform printed on a chart paper. An automated disease classification and prediction is yet to become popular and such a feature is found in very few ECG machine.

In my project an attempt was made to arrive at an automated disease classification and prediction using ECG signal and machine learning concept like K Nearest Neighbour (KNN). The focus was restrained to Arrhythmia prediction. The approach adopted is to detect QRS complex based on KNN. Calculate Heart rate from the QRS complexes. A clinical standard called Nova code was adopted which provides quantitative information regarding Arrhythmia parameters. Further the false positives(FPs) and false negatives (FNs) obtained during QRS detection was rectified with a custom designed Fn and FP correction algorithm. Further a P detection algorithm was developed. Finally disease classification was obtained with a combination of KNN and Fuzzy logic.